Add files via upload
It is a simulation data for a survey conducted by an imaginary company ‘’ÄLPHABETAGAMA’’ for it  customers. There are 7 questions Q1 to Q7. They are all categorical variables. Rest are all integers. Based on the responses the company classified their customers Risk Type as “High”and “Low”. Chossing this Risk type as dependent variable and creating a prediction model.
Note : This is an Unbalance data